rename from libKitsunemimiProjectNetwork to libKitsunemimiSakuraNetwork
