+++ include/libKitsunemimiSakuraNetwork/session.h
