+++ src/messages_processing/error_processing.h
@@ -32,8 +32,8 @@
 #include <libKitsunemimiNetwork/abstract_socket.h>
 #include <libKitsunemimiCommon/buffer/ring_buffer.h>
 
-#include <libKitsunemimiProjectNetwork/session_controller.h>
-#include <libKitsunemimiProjectNetwork/session.h>
+#include <libKitsunemimiSakuraNetwork/session_controller.h>
+#include <libKitsunemimiSakuraNetwork/session.h>
 
 #include <libKitsunemimiPersistence/logger/logger.h>
