+++ src/messages_processing/stream_data_processing.h
@@ -30,8 +30,8 @@
 #include <libKitsunemimiNetwork/abstract_socket.h>
 #include <libKitsunemimiCommon/buffer/ring_buffer.h>
 
-#include <libKitsunemimiProjectNetwork/session_controller.h>
-#include <libKitsunemimiProjectNetwork/session.h>
+#include <libKitsunemimiSakuraNetwork/session_controller.h>
+#include <libKitsunemimiSakuraNetwork/session.h>
 
 #include <libKitsunemimiPersistence/logger/logger.h>
