+++ tests/benchmark_tests/test_session.cpp
@@ -2,8 +2,8 @@
 
 #include <libKitsunemimiCommon/buffer/stack_buffer.h>
 
-#include <libKitsunemimiProjectNetwork/session.h>
-#include <libKitsunemimiProjectNetwork/session_controller.h>
+#include <libKitsunemimiSakuraNetwork/session.h>
+#include <libKitsunemimiSakuraNetwork/session_controller.h>
 
 #include <libKitsunemimiCommon/common_methods/string_methods.h>
 #include <libKitsunemimiCommon/buffer/data_buffer.h>
