+++ tests/cli_tests/test_session.cpp
@@ -1,7 +1,7 @@
 ﻿#include "test_session.h"
 
-#include <libKitsunemimiProjectNetwork/session.h>
-#include <libKitsunemimiProjectNetwork/session_controller.h>
+#include <libKitsunemimiSakuraNetwork/session.h>
+#include <libKitsunemimiSakuraNetwork/session_controller.h>
 #include <libKitsunemimiCommon/common_methods/string_methods.h>
 #include <libKitsunemimiCommon/buffer/data_buffer.h>
